refactor(plugins): promote plugins to internal/plugins
Move-only: the plugin installer, manifest, source parser and the
HookCommandProvider that satisfies agentdomain.HookCommandProvider. This edge
is one of the two that made internal/agent import upward - internal/agent now
imports internal/plugins, a sibling context, instead.
