+++ internal/plugins/hooks.go
