+++ internal/plugins/hooks_test.go
