+++ internal/plugins/installer.go
