+++ internal/plugins/installer_test.go
