+++ internal/plugins/instructions_test.go
