+++ internal/plugins/manifest.go
